+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/PurchaseOrderInvoiceProjectServiceImpl.java
@@ -98,7 +98,7 @@ public class PurchaseOrderInvoiceProjectServiceImpl extends PurchaseOrderInvoice
 
 			InvoiceLineGenerator invoiceLineGenerator = new InvoiceLineGenerator(invoice, product, product.getName(), price,
 						price,purchaseOrderLine.getDescription(),purchaseOrderLine.getQty(),purchaseOrderLine.getUnit(), null,InvoiceLineGenerator.DEFAULT_SEQUENCE,discountAmount,discountTypeSelect,
-						null, null,false)  {
+						null, null,false, false)  {
 				@Override
 				public List<InvoiceLine> creates() throws AxelorException {
 
@@ -116,7 +116,7 @@ public class PurchaseOrderInvoiceProjectServiceImpl extends PurchaseOrderInvoice
 			price = price.multiply(invoice.getPartner().getChargeBackPurchase().divide(new BigDecimal(100), appBusinessProjectService.getNbDecimalDigitForUnitPrice(), BigDecimal.ROUND_HALF_UP)).setScale(appBusinessProjectService.getNbDecimalDigitForUnitPrice(), BigDecimal.ROUND_HALF_UP);
 			InvoiceLineGenerator invoiceLineGenerator = new InvoiceLineGenerator(invoice, product, product.getName(), price,
 						price,purchaseOrderLine.getDescription(),purchaseOrderLine.getQty(),purchaseOrderLine.getUnit(), null,InvoiceLineGenerator.DEFAULT_SEQUENCE,discountAmount,discountTypeSelect,
-						null, null,false)  {
+						null, null,false, false)  {
 				@Override
 				public List<InvoiceLine> creates() throws AxelorException {
 
@@ -134,7 +134,7 @@ public class PurchaseOrderInvoiceProjectServiceImpl extends PurchaseOrderInvoice
 		else{
 			InvoiceLineGeneratorSupplyChain invoiceLineGenerator = new InvoiceLineGeneratorSupplyChain(invoice, product, purchaseOrderLine.getProductName(),
 					purchaseOrderLine.getDescription(), purchaseOrderLine.getQty(), purchaseOrderLine.getUnit(),
-					purchaseOrderLine.getSequence(), false, null, purchaseOrderLine, null)  {
+					purchaseOrderLine.getSequence(), false, null, purchaseOrderLine, null, false)  {
 				@Override
 				public List<InvoiceLine> creates() throws AxelorException {