+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/SaleOrderServiceImpl.java
@@ -17,6 +17,21 @@
  */
 package com.axelor.apps.sale.service;
 
+import java.lang.invoke.MethodHandles;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.Iterator;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import javax.persistence.Query;
+import com.axelor.apps.base.db.repo.PriceListRepository;
+import com.axelor.apps.base.service.PartnerPriceListService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import com.axelor.apps.ReportFactory;
 import com.axelor.apps.base.db.AppSale;
 import com.axelor.apps.base.db.CancelReason;
@@ -59,7 +74,6 @@ import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import javax.persistence.Query;
 import java.lang.invoke.MethodHandles;
 import java.math.BigDecimal;
@@ -545,21 +559,25 @@ public class SaleOrderServiceImpl implements SaleOrderService {
 	private void _addPackLines(SaleOrder saleOrder) {
 		
 		if (saleOrder.getSaleOrderLineList() == null) {
-			return;
-		}
-		
-		List<SaleOrderLine> lines = new ArrayList<SaleOrderLine>();
-		lines.addAll(saleOrder.getSaleOrderLineList());
-		for (SaleOrderLine line : lines) {
-			if (line.getSubLineList() == null) {
-				continue;
-			}
-			for (SaleOrderLine subLine : line.getSubLineList()) {
-				if (subLine.getSaleOrder() == null) {
-					saleOrder.addSaleOrderLineListItem(subLine);
-				}
-			}
-		}
+            return;
+        }
+        
+        List<SaleOrderLine> lines = new ArrayList<SaleOrderLine>();
+        lines.addAll(saleOrder.getSaleOrderLineList());
+        for (SaleOrderLine line : lines) {
+            if (line.getSubLineList() == null || line.getSubLineList().isEmpty()) {
+                continue;
+            }
+            int subSequence = 1;
+            for (SaleOrderLine subLine : line.getSubLineList()) {
+                if (subLine.getSaleOrder() == null) {
+                	subLine.setSequence(subSequence + line.getSequence());
+                    saleOrder.getSaleOrderLineList().add(subLine);
+                    subSequence++;
+                }
+            }
+            
+        }
 	}
 
 	@Override
@@ -608,5 +626,34 @@ public class SaleOrderServiceImpl implements SaleOrderService {
             saleOrder.getSaleOrderLineList().sort(Comparator.comparing(SaleOrderLine::getSequence));
         }
     }
+    
+	public List<SaleOrderLine> removeSubLines(List<SaleOrderLine> soLines) {
 
+		if (soLines == null) {
+	        return soLines;
+	    }
+	    
+	    
+	    List<SaleOrderLine> subLines = new ArrayList<SaleOrderLine>();
+	    for (SaleOrderLine packLine : soLines) {
+	        if (packLine.getTypeSelect() == 2 && packLine.getSubLineList() != null) {
+	            packLine.getSubLineList().removeIf(it->it.getId() != null && !soLines.contains(it));
+	            packLine.setTotalPack(packLine.getSubLineList()
+	                    .stream().map(it->it.getExTaxTotal()).reduce(BigDecimal.ZERO, BigDecimal::add));
+	            subLines.addAll(packLine.getSubLineList());
+	        }
+	    }
+	    Iterator<SaleOrderLine> lines = soLines.iterator();
+
+	    while (lines.hasNext()) {
+	        SaleOrderLine subLine = lines.next();
+	        if (subLine.getId() != null 
+	                && subLine.getParentLine() != null
+	                && !subLines.contains(subLine)) {
+	                lines.remove();
+	        }
+	    }
+
+	    return soLines;
+	}
 }