+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/invoice/InvoiceServiceSupplychainImpl.java
@@ -17,17 +17,16 @@
  */
 package com.axelor.apps.supplychain.service.invoice;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
-
 import com.axelor.apps.account.db.Invoice;
 import com.axelor.apps.account.db.InvoiceLine;
-import com.axelor.apps.account.db.Move;
 import com.axelor.apps.account.db.MoveLine;
 import com.axelor.apps.account.db.repo.InvoiceRepository;
 import com.axelor.apps.account.service.app.AppAccountService;
@@ -41,12 +40,13 @@ import com.axelor.apps.base.db.Currency;
 import com.axelor.apps.base.service.alarm.AlarmEngineService;
 import com.axelor.apps.sale.db.AdvancePayment;
 import com.axelor.apps.sale.db.SaleOrder;
+import com.axelor.apps.sale.db.SaleOrderLine;
 import com.axelor.db.Query;
 import com.axelor.exception.AxelorException;
 import com.axelor.inject.Beans;
 import com.google.inject.Inject;
 
-public class InvoiceServiceSupplychainImpl extends InvoiceServiceImpl {
+public class InvoiceServiceSupplychainImpl extends InvoiceServiceImpl implements InvoiceServiceSupplychain {
 
 	@Inject
 	public InvoiceServiceSupplychainImpl(ValidateFactory validateFactory, VentilateFactory ventilateFactory,
@@ -127,5 +127,58 @@ public class InvoiceServiceSupplychainImpl extends InvoiceServiceImpl {
 					.flatMap(move -> move.getMoveLineList().stream())
 					.collect(Collectors.toList());
 		}
+	}
+	
+	@Override
+	public List<InvoiceLine> addSubLines(List<InvoiceLine> invoiceLine) {
+		
+		if (invoiceLine == null) {
+            return invoiceLine;
+        }
+        
+        List<InvoiceLine> lines = new ArrayList<InvoiceLine>();
+        lines.addAll(invoiceLine);
+        for (InvoiceLine line : lines) {
+            if (line.getSubLineList() == null) {
+                continue;
+            }
+            for (InvoiceLine subLine : line.getSubLineList()) {
+                if (subLine.getInvoice() == null) {
+                	invoiceLine.add(subLine);
+                }
+            }
+        }
+		return invoiceLine;
+	}
+	
+	@Override
+	public List<InvoiceLine> removeSubLines(List<InvoiceLine> invoiceLines) {
+		
+		if (invoiceLines == null) {
+	        return invoiceLines;
+	    }
+	    
+	    
+	    List<InvoiceLine> subLines = new ArrayList<InvoiceLine>();
+	    for (InvoiceLine packLine : invoiceLines) {
+	        if (packLine.getTypeSelect() == 2 && packLine.getSubLineList() != null) {
+	            packLine.getSubLineList().removeIf(it->it.getId() != null && !invoiceLines.contains(it));
+	            packLine.setTotalPack(packLine.getSubLineList()
+	                    .stream().map(it->it.getExTaxTotal()).reduce(BigDecimal.ZERO, BigDecimal::add));
+	            subLines.addAll(packLine.getSubLineList());
+	        }
+	    }
+	    Iterator<InvoiceLine> lines = invoiceLines.iterator();
+
+	    while (lines.hasNext()) {
+	    	InvoiceLine subLine = lines.next();
+	        if (subLine.getId() != null 
+	                && subLine.getParentLine() != null
+	                && !subLines.contains(subLine)) {
+	                lines.remove();
+	        }
+	    }
+
+	    return invoiceLines;
 	}
 }