+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/InvoiceController.java
@@ -33,8 +33,10 @@ import com.axelor.apps.base.db.Wizard;
 import com.axelor.apps.sale.db.SaleOrder;
 import com.axelor.apps.supplychain.service.SaleOrderInvoiceService;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
+import com.axelor.apps.supplychain.service.invoice.InvoiceServiceSupplychain;
 import com.axelor.db.JPA;
 import com.axelor.exception.AxelorException;
+import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
 import com.axelor.meta.schema.actions.ActionView;
 import com.axelor.meta.schema.actions.ActionView.ActionViewBuilder;
@@ -50,6 +52,9 @@ public class InvoiceController {
 	
 	@Inject
 	protected SaleOrderInvoiceService saleOrderInvoiceService;
+	
+	@Inject
+	private InvoiceServiceSupplychain invoiceService;
 	
 	public void fillInLines(ActionRequest request, ActionResponse response) {
 		Invoice invoice = request.getContext().asType(Invoice.class);
@@ -258,4 +263,24 @@ public class InvoiceController {
 				response.setFlash(ae.getLocalizedMessage());
 			}
 		}
+		
+		public void addSubLines(ActionRequest request, ActionResponse response) {
+	    	try {
+	            Invoice invoice = request.getContext().asType(Invoice.class);
+	            response.setValue("invoiceLineList",  invoiceService.addSubLines(invoice.getInvoiceLineList()));
+	        } catch (Exception e) {
+	            TraceBackService.trace(response, e);
+	            response.setReload(true);
+	        }
+	    }
+		
+		public void removeSubLines(ActionRequest request, ActionResponse response) {
+	        try {
+	            Invoice invoice = request.getContext().asType(Invoice.class);
+	            response.setValue("invoiceLineList",  invoiceService.removeSubLines(invoice.getInvoiceLineList()));
+	        } catch (Exception e) {
+	            TraceBackService.trace(response, e);
+	            response.setReload(true);
+	        }
+	    }
 }